Make a wrapper class private
